Remove zero on hand inventory records, closes #10

+++ src/main/java/skim/sfg/beer/inventory/service/services/AllocationServiceImpl.java
@@ -74,6 +74,9 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLineDto.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
 
+            }
+
+            if (beerInventory.getQuantityOnHand() == 0) {
                 beerInventoryRepository.delete(beerInventory);
             }
         });
